Create README.md
I am interested in coding and l also work with organisations like ZINDI they use large models and big data to solve Africa critical and pressing issues. Girls in tech Zimbabwe is one of the organisations that l engage with to educate women so that they equip tech skills so that they will be  relevant in the future.